+++ libs/opengl/include/mrpt/opengl/CColorBar.h
@@ -52,7 +52,8 @@ class CColorBar : public CRenderizableShaderTriangles,
 	virtual shader_list_t requiredShaders() const override
 	{
 		// May use up to two shaders (triangles and lines):
-		return {DefaultShaderID::WIREFRAME, DefaultShaderID::TRIANGLES};
+		return {
+			DefaultShaderID::WIREFRAME, DefaultShaderID::TRIANGLES_NO_LIGHT};
 	}
 	void onUpdateBuffers_Wireframe() override;
 	void onUpdateBuffers_Triangles() override;
@@ -64,7 +65,7 @@ class CColorBar : public CRenderizableShaderTriangles,
 	}
 	/** @} */
 
-	mrpt::math::TBoundingBox getBoundingBox() const override;
+	mrpt::math::TBoundingBoxf internalBoundingBoxLocal() const override;
 
 	void setColormap(const mrpt::img::TColormap colormap);
 	void setColorAndValueLimits(