+++ libs/opengl/include/mrpt/opengl/CSetOfTexturedTriangles.h
@@ -33,7 +33,7 @@ class CSetOfTexturedTriangles : public CRenderizableShaderTexturedTriangles
 
 	/** Evaluates the bounding box of this object (including possible children)
 	 * in the coordinate frame of the object parent. */
-	mrpt::math::TBoundingBox getBoundingBox() const override;
+	mrpt::math::TBoundingBoxf internalBoundingBoxLocal() const override;
 
 	void clearTriangles()
 	{