+++ libs/opengl/src/CColorBar.cpp
@@ -67,7 +67,7 @@ void CColorBar::render(const RenderContext& rc) const
 	// colobars are typically displayed on-top
 	switch (rc.shader_id)
 	{
-		case DefaultShaderID::TRIANGLES:
+		case DefaultShaderID::TRIANGLES_NO_LIGHT:
 			CRenderizableShaderTriangles::render(rc);
 			break;
 		case DefaultShaderID::WIREFRAME:
@@ -231,8 +231,7 @@ void CColorBar::serializeFrom(
 	CRenderizable::notifyChange();
 }
 
-auto CColorBar::getBoundingBox() const -> mrpt::math::TBoundingBox
+mrpt::math::TBoundingBoxf CColorBar::internalBoundingBoxLocal() const
 {
-	return mrpt::math::TBoundingBox({0, 0, 0}, {m_width, m_height, .0})
-		.compose(m_pose);
+	return {{.0f, .0f, .0f}, {d2f(m_width), d2f(m_height), .0f}};
 }