+++ libs/opengl/src/CRenderizableShaderPoints.cpp
@@ -89,7 +89,7 @@ void CRenderizableShaderPoints::render(const RenderContext& rc) const
 			0, /* stride */
 			BUFFER_OFFSET(0) /* array buffer offset */
 		);
-		CHECK_OPENGL_ERROR();
+		CHECK_OPENGL_ERROR_IN_DEBUG();
 	}
 
 	// Set up the color array:
@@ -107,15 +107,15 @@ void CRenderizableShaderPoints::render(const RenderContext& rc) const
 			0, /* stride */
 			BUFFER_OFFSET(0) /* array buffer offset */
 		);
-		CHECK_OPENGL_ERROR();
+		CHECK_OPENGL_ERROR_IN_DEBUG();
 	}
 
 	glDrawArrays(GL_POINTS, 0, m_vertex_buffer_data.size());
-	CHECK_OPENGL_ERROR();
+	CHECK_OPENGL_ERROR_IN_DEBUG();
 
 	if (attr_position) glDisableVertexAttribArray(*attr_position);
 	if (attr_color) glDisableVertexAttribArray(*attr_color);
-	CHECK_OPENGL_ERROR();
+	CHECK_OPENGL_ERROR_IN_DEBUG();
 #endif
 }
 
@@ -141,23 +141,17 @@ void CRenderizableShaderPoints::params_deserialize(
 	};
 }
 
-const mrpt::math::TBoundingBox CRenderizableShaderPoints::verticesBoundingBox()
+const mrpt::math::TBoundingBoxf CRenderizableShaderPoints::verticesBoundingBox()
 	const
 {
 	std::shared_lock<std::shared_mutex> wfReadLock(
 		CRenderizableShaderPoints::m_pointsMtx.data);
 
-	mrpt::math::TBoundingBox bb;
+	mrpt::math::TBoundingBoxf bb;
 
 	if (m_vertex_buffer_data.empty()) return bb;
 
-	bb.min = mrpt::math::TPoint3D(
-		std::numeric_limits<double>::max(), std::numeric_limits<double>::max(),
-		std::numeric_limits<double>::max());
-	bb.max = mrpt::math::TPoint3D(
-		-std::numeric_limits<double>::max(),
-		-std::numeric_limits<double>::max(),
-		-std::numeric_limits<double>::max());
+	bb = mrpt::math::TBoundingBoxf::PlusMinusInfinity();
 
 	for (const auto& p : m_vertex_buffer_data)
 	{